+++ packages/cursorless-engine/src/scripts/transformRecordedTests/checkMarks.ts
@@ -3,7 +3,7 @@ import { uniq } from "lodash";
 import { injectIde } from "../../singletons/ide.singleton";
 import tokenGraphemeSplitter from "../../singletons/tokenGraphemeSplitter.singleton";
 import { extractTargetKeys } from "../../testUtil/extractTargetKeys";
-import { getPartialTargetDescriptors } from "../../util/getPartialTargetDescriptors.1";
+import { getPartialTargetDescriptors } from "../../util/getPartialTargetDescriptors";
 import { upgrade } from "./transformations/upgrade";
 import assert = require("assert");
