+++ packages/cursorless-engine/src/util/getPartialTargetDescriptors.ts
